# PERF: DataFrame(ndarr, copy=True)