--- a/pandas/core/internals/construction.py
+++ b/pandas/core/internals/construction.py
@@ -312,8 +312,7 @@ def ndarray_to_mgr(
             if (dtype is None or astype_is_view(values.dtype, dtype))
             else False
         )
-        _order = "F" if _copy else "K"
-        values = np.array(values, copy=_copy, order=_order)
+        values = np.array(values, copy=_copy, order=("F" if _copy else "K"))
         values = _ensure_2d(values)
 
     else: