@phofl 2023-04-06
Not sure I've ever ran the groupby ones, the frame for example take around 5-10 minutes (not totally sure, would lean more towards 5) on my m2 air