@topper-123 2023-04-06
So `asv continuous -f 1.1 upstream/main HEAD -b frame_methods` takes approx 5 min. in for you?

Did you install asv like in the [docs](https://pandas.pydata.org/docs/development/contributing_codebase.html#running-the-performance-test-suite) i.e. (`pip install git+https://github.com/airspeed-velocity/asv`) or did you use conda or pip?

(I need to try figuring it out, so appreciate any help)