@phofl 2023-04-06
I think so yes, can re-run later today. It's definitely not terribly long

I used mamba (came when installing environment.yml)