@phofl 2023-04-06
5 was a bit optimistic, ran around 9ish minutes. 

2 suggestions:

- set the number of cores when compiling in asv.conf.json to the number of cores that your machine has (speeds it up quite nicely).
- check that you are not on low power mode (battery settings). test suite takes 2.5 times as long for me when on low power mode