@topper-123 2023-04-06
Great, thanks I will try that, alternatively do a new install using mamba.