@topper-123 2023-04-12
Hmm, setting the default `copy=True`  and comparing against that will do a lot of other things also. In  particular, every time `DataFrame._constructor` & `Series._constructor_expanddim` are called, a copy will be made, which will be quite costly. That isn't proposed in this PR, so is not really the same.

This PR doesn't do anything that `DataFrame.copy` doesn't do already, so will logically have comparable performance to dataframes that come out of `DataFrame.copy()`. So I don't see any possibility that e.g. arithmetic operations can be affected negatively from this.

I could add a test that does `pd.DataFrame(data, copy=True/False).sum()` somewhere to in the ASVs to document the case where this gets a performance boost.